// web/src/api/models/chat-details.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ATRAG API
 * ATRAG API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ChatMessage } from './chat-message';

/**
 * 
 * @export
 * @interface ChatDetails
 */
export interface ChatDetails {
    /**
     * 
     * @type {string}
     * @memberof ChatDetails
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof ChatDetails
     */
    'title'?: string;
    /**
     * 
     * @type {string}
     * @memberof ChatDetails
     */
    'bot_id'?: string;
    /**
     * 
     * @type {string}
     * @memberof ChatDetails
     */
    'peer_id'?: string;
    /**
     * 
     * @type {string}
     * @memberof ChatDetails
     */
    'peer_type'?: ChatDetailsPeerTypeEnum;
    /**
     * Array of conversation turns, where each turn is an array of message parts
     * @type {Array<Array<ChatMessage>>}
     * @memberof ChatDetails
     */
    'history'?: Array<Array<ChatMessage>>;
    /**
     * 
     * @type {string}
     * @memberof ChatDetails
     */
    'status'?: ChatDetailsStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof ChatDetails
     */
    'created'?: string;
    /**
     * 
     * @type {string}
     * @memberof ChatDetails
     */
    'updated'?: string;
}

export const ChatDetailsPeerTypeEnum = {
    system: 'system',
    feishu: 'feishu',
    weixin: 'weixin',
    weixin_official: 'weixin_official',
    web: 'web',
    dingtalk: 'dingtalk'
} as const;

export type ChatDetailsPeerTypeEnum = typeof ChatDetailsPeerTypeEnum[keyof typeof ChatDetailsPeerTypeEnum];
export const ChatDetailsStatusEnum = {
    active: 'active',
    archived: 'archived'
} as const;

export type ChatDetailsStatusEnum = typeof ChatDetailsStatusEnum[keyof typeof ChatDetailsStatusEnum];